Fix : Bug Total gaji: Int -> Double

+++ app/src/main/java/id/my/matahati/absensi/data/SalaryComponent.kt
@@ -2,5 +2,5 @@ package id.my.matahati.absensi.data
 
 data class SalaryComponent(
     val label: String,
-    val value: Int
+    val value: Double
 )
